+++ src/PluginInfo.cs
@@ -4,5 +4,5 @@ public static class PluginInfo
 {
 	public const string PLUGIN_GUID = "moe.sylvi.WeekdayBonuses";
 	public const string PLUGIN_NAME = "WeekdayBonuses";
-	public const string PLUGIN_VERSION = "1.1.0";
+	public const string PLUGIN_VERSION = "1.1.1";
 }