+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-### v1.1.0
+## v1.1.1
+- Add Double Monster config for when the guaranteed Old Bird wakes up
+- Add landmine modifiers to Double Traps
+    - Configurable delay for the landmine to explode after stepping off of it
+    - Configurable damage for the non-lethal radius of the landmine explosion
+
+## v1.1.0
 - Add Double Monster Monday
 - Add Double Trap Tuesday
 - Move Small Facility to Thursday
@@ -7,5 +13,5 @@
 - Add config for Small Facility to apply a multiplier
 - Add config to have multiple events on one day
 
-### v1.0.0
+## v1.0.0
 - Initial release